work in progress on wdf report
